% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartDiag.R
\name{bartClassifDiag}
\alias{bartClassifDiag}
\title{bartClassifDiag}
\usage{
bartClassifDiag(
  model,
  data,
  response,
  threshold = "Youden",
  pNorm = FALSE,
  showInterval = TRUE,
  combineFact = FALSE
)
}
\arguments{
\item{model}{a model created from either the BART, dbarts, or bartMachine package.}

\item{data}{A dataframe}

\item{response}{The name of the response for the fit.}

\item{threshold}{A dashed line on some plots to indicate a chosen threshold value.
by default the Youden index is shown.}

\item{pNorm}{apply pnorm to the y-hat data}

\item{showInterval}{LOGICAL if TRUE then show 5\% and 95\% quantile intervals.}
}
\value{
A selection of diagnostic plots
}
\description{
Displays a selection of diagnostic plots for a BART model.
}
