% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartMachineTreeData.R
\name{bartMachineTreeData}
\alias{bartMachineTreeData}
\title{bartMachineTreeData}
\usage{
bartMachineTreeData(model)
}
\arguments{
\item{model}{Model created from the bartMachine package.}
}
\value{
A data frame of every tree and its attributes.
}
\description{
Creates a data frame of all tree attributes for a model
created by the bartMachine package.
}
