% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartMachineTreeList.R
\name{bartMachineTreeList}
\alias{bartMachineTreeList}
\title{bartMachineTreeList}
\usage{
bartMachineTreeList(trees)
}
\arguments{
\item{trees}{A data frame created by bartMachineTreeData function.}
}
\value{
A list of every tree and its attributes.
}
\description{
Creates a list of tree attributes for a model
created by the bartMachine package.
}
