% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartTreesData.R
\name{bartTreeData}
\alias{bartTreeData}
\title{bartTreeData}
\usage{
bartTreeData(model)
}
\arguments{
\item{model}{Model created from the bart package.}
}
\value{
A data frame of every tree and its attributes.
}
\description{
Creates a data frame of all tree attributes for a model
created by the bart package.
}
