% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartTreesList.R
\name{bartTreeList}
\alias{bartTreeList}
\title{bartTreeList}
\usage{
bartTreeList(trees)
}
\arguments{
\item{trees}{A data frame created by bartTreesData function.}
}
\value{
A list of every tree and its attributes.
}
\description{
Creates a list of tree attributes for a model
created by the bart package.
}
