% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{clusterTrees}
\alias{clusterTrees}
\title{Cluster Trees by Variable}
\usage{
clusterTrees(tree_list)
}
\arguments{
\item{tree_list}{A list of trees, where each tree is expected to have a 'var' column.}
}
\value{
A list of trees reordered based on the clustering of variables.
}
\description{
Reorders a list of tree structures based on the clustering of variables within each tree.
}
