% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbartsTreeData.R
\name{dbartsTreeData}
\alias{dbartsTreeData}
\title{dbartsTreeData}
\usage{
dbartsTreeData(model)
}
\arguments{
\item{model}{Model created from the dbarts package.}
}
\value{
A data frame of every tree and its attributes.
}
\description{
Creates a data frame of all tree attributes for a model
created by the dbarts package.
}
