% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbartsTreeList.R
\name{dbartsTreeList}
\alias{dbartsTreeList}
\title{dbartsTreeList}
\usage{
dbartsTreeList(trees)
}
\arguments{
\item{trees}{A data frame created by dbartsTreeData function.}
}
\value{
A list of every tree and its attributes.
}
\description{
Creates a list of tree attributes for a model
created by the dbarts package.
}
