% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_stump_colour_for_legend}
\alias{get_stump_colour_for_legend}
\title{Determines the stump color for a legend based on its mean value}
\usage{
get_stump_colour_for_legend(lims, mean_value, palette)
}
\arguments{
\item{lims}{A numeric vector of length 2 specifying the limits within which the mean value falls.}

\item{mean_value}{The mean value for which the color needs to be determined.}

\item{palette}{A character vector of colors representing the palette from which the color is selected.}
}
\value{
A character string specifying the color corresponding to the mean value.
}
\description{
This function is internal and is used to compute the color of a stump
for the purpose of legend display, based on the mean value relative to specified limits.
}
