% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximityMatrix.R
\name{getIndex}
\alias{getIndex}
\title{getIndex}
\usage{
getIndex(tree, data, nRows)
}
\arguments{
\item{data}{Data frame of data excluding the response.}

\item{nRows}{Sequence of integers from 1 to the number of rows in the data.}

\item{treeData}{A data frame created by treeData function.}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
gets the indices of observations from leaf nodes.
}
