% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeObsIndex.R
\name{getNodeObsIndex}
\alias{getNodeObsIndex}
\title{getNodeObsIndex}
\usage{
getNodeObsIndex(tree, var, nRows)
}
\arguments{
\item{nRows}{Sequence of integers from 1 to the number of rows in the data.}

\item{treeData}{A data frame created by treeData function.}

\item{data}{Data frame of data excluding the response.}
}
\value{
A list containing node attributes.
}
\description{
Gets the index of the observations contained within a node.
}
