% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{getObservations}
\alias{getObservations}
\title{Get Observations Falling into Each Node}
\usage{
getObservations(data, treeData)
}
\arguments{
\item{data}{A data frame used to build BART model.}

\item{treeData}{A data frame representing the tree structure, including the
necessary columns `iteration`, `treeNum`, `var`, and `splitValue`.}
}
\value{
A modified version of `treeData` that includes two new columns: `obsNode` and
        `noObs`. `obsNode` lists the observations falling into each node, and
        `noObs` provides the count of observations for each node.
}
\description{
This function determines which observations from a given dataset fall into
which nodes of a tree, based on a tree structure defined in `treeData`.
The treeData object must include `iteration`, `treeNum`, `var`, and `splitValue` columns.
}
\examples{
 \donttest{
treeDataWithObservations <- getObservations(data = my_data, treeData = my_trees)
}
}
