% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSUP.R
\name{guide_colourfan}
\alias{guide_colourfan}
\alias{guide_colorfan}
\title{Colourfan guide}
\usage{
guide_colourfan(
  title = waiver(),
  title.x.position = "top",
  title.y.position = "right",
  title.theme = NULL,
  title.hjust = 0.5,
  title.vjust = NULL,
  label = TRUE,
  label.theme = NULL,
  barwidth = NULL,
  barheight = NULL,
  nbin = 32,
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill"),
  ...
)

guide_colorfan(
  title = waiver(),
  title.x.position = "top",
  title.y.position = "right",
  title.theme = NULL,
  title.hjust = 0.5,
  title.vjust = NULL,
  label = TRUE,
  label.theme = NULL,
  barwidth = NULL,
  barheight = NULL,
  nbin = 32,
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill"),
  ...
)
}
\arguments{
\item{title}{Title}

\item{title.x.position}{Title x position}

\item{title.y.position}{Title y position}

\item{title.theme}{Title theme}

\item{title.hjust}{Title hjust}

\item{title.vjust}{Title vjust}

\item{label}{Label}

\item{label.theme}{Label theme}

\item{barwidth}{Barwidth}

\item{barheight}{Barheight}

\item{nbin}{Number of bins}

\item{reverse}{Reverse}

\item{order}{order}

\item{available_aes}{Available aesthetics}

\item{...}{Extra paramters}
}
\value{
A `grob` object representing a color fan. This `grob` can be added to a grid-based plot
or a ggplot2 object to visualize a range of colors in a fan-like structure. Each segment of the fan
corresponds to a color specified in the `colours` parameter, allowing for an intuitive representation
of color gradients or palettes.
}
\description{
Colourfan guide
}
