% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{input_data}
\alias{input_data}
\title{input_data}
\format{
A data frame with 10 rows and 6 columns:
\describe{
  \item{x1}{Covariate}
  \item{x2}{Covariate}
  \item{x3}{Covariate}
  \item{x4}{Covariate}
  \item{x5}{Covariate}
  \item{y}{Response}
  ...
}
}
\usage{
input_data
}
\description{
Small example of Friedman data following the formula:
\deqn{y = 10 \sin(\pi x_1 x_2) + 20 (x_3 - 0.5)^2 + 10 x_4 + 5 x_5 + e}
}
\keyword{datasets}
