% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOfTrees.R
\name{listOfTrees}
\alias{listOfTrees}
\title{listOfTrees}
\usage{
listOfTrees(treeData)
}
\arguments{
\item{treeData}{A list of tree attributes created by extractTreeData function.}
}
\value{
A list of every tree and its attributes.
}
\description{
Creates a list of tree attributes for a model
created by either the BART, dbarts, or bartMachine packages.
}
