% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permBART.R
\name{permBART}
\alias{permBART}
\title{permBART}
\usage{
permBART(model, data, numTreesPerm = NULL, plotType = "barplot")
}
\arguments{
\item{model}{Model created from either the BART, dbarts or bartMachine packages.}

\item{data}{A data frame containing variables in the model.}

\item{numTreesPerm}{The number of trees to be used in the null model.
As suggested by Chipman (2009), a small number of trees is recommended (~20) to force important
variables to used in the model. If NULL, then the number of trees from the true model is used.}
}
\value{
A variable selection plot.
}
\description{
A variable selection approach which creates a null model by
permuting the response, rebuilding the model, and calculating the inclusion proportion (IP) on the null model.
The final result displayed is the original model's IP  minus the null IP.
}
