% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTreesOLD.R
\name{plotAllTreesOLD}
\alias{plotAllTreesOLD}
\title{plotAllTreesOLD}
\usage{
plotAllTreesOLD(treeList, sampleSize = 0, cluster = FALSE)
}
\arguments{
\item{treeList}{A list of trees created by treeList function.}

\item{sampleSize}{Sample the tree list.}

\item{cluster}{LOGICAL. If TRUE, then cluster by tree structures.}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
Plots all the trees created by the treeList function.
}
