% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{plotMDS}
\usage{
plotMDS(matrix, pal = NULL, response)
}
\arguments{
\item{matrix}{A matrix of proximities created by the proximityMatrix function}

\item{pal}{A vector of colours used to colour the facrtor levels..}

\item{response}{The respose used to get factor levels to colour points in plot.}
}
\value{
A MDS plot.
}
\description{
Multi-dimensional Scaling Plot of proximity matrix from a BART model.
}
