% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(treeData, iter = 1, treeNo = 1, plotType = c("dendrogram", "icicle"))
}
\arguments{
\item{treeData}{A data frame created by treeData function}

\item{iter}{The MCMC iteration or chain to plot.}

\item{treeNo}{The tree number to plot.}

\item{plotType}{What type of plot to display. either dendrogram or icicle.}
}
\value{
A plot of an individual tree
}
\description{
plots individual trees from the R-packages BART, dbarts, and
bartMachine
}
