% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTreeData.R
\name{print.hideHelper}
\alias{print.hideHelper}
\title{print.hideHelper}
\usage{
\method{print}{hideHelper}(x, ...)
}
\arguments{
\item{x}{A data frame of trees}

\item{...}{Extra parameters}
}
\description{
This function hides some columns from the print out
but are still accessible via indexing.
}
