% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximityMatrixOLD.R
\name{proximityMatrixOLD}
\alias{proximityMatrixOLD}
\title{proximityMatrixOLD}
\usage{
proximityMatrixOLD(tree, data, nRows, normalize = TRUE, reorder = TRUE)
}
\arguments{
\item{data}{Data frame of data excluding the response.}

\item{nRows}{Number of rows to consider.}

\item{normalize}{Default is TRUE. Divide the total number of pairs of observations by
the number of trees.}

\item{reorder}{Default is TRUE. Whether to sort the matrix so high values are pushed to top left.}

\item{treeData}{A list of tree attributes created by treeData function.}
}
\value{
A matrix containing proximity values.
}
\description{
Creates a matrix of proximity values.
}
