% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sort_trees_by_depthMax}
\alias{sort_trees_by_depthMax}
\title{Sort Trees by Maximum Depth}
\usage{
sort_trees_by_depthMax(tree_list)
}
\arguments{
\item{tree_list}{List of `tbl_graph` trees.}
}
\value{
Sorted list of `tbl_graph` trees by decreasing maximum depth.
}
\description{
Sort Trees by Maximum Depth
}
