% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{terminalFunction}
\alias{terminalFunction}
\title{Generate Terminal Node Indicator}
\usage{
terminalFunction(data)
}
\arguments{
\item{data}{A data frame containing tree structure information with at least `treeNum`,
`iteration`, and `depth` columns.}
}
\value{
The modified data frame with an additional `terminal` column.
}
\description{
Adds a boolean `terminal` column to the dataset indicating whether each node is terminal.
}
