% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSUP.R
\name{train_bivariate}
\alias{train_bivariate}
\title{Train range for bivariate scale}
\usage{
train_bivariate(new, existing = NULL)
}
\arguments{
\item{new}{New data on which to train.}

\item{existing}{Existing range}
}
\value{
A tibble containing two columns, `range1` and `range2`, each representing the trained
continuous range based on the new and existing data. This function is used to update or define
the scales of a bivariate analysis by considering both new input data and any existing range
specifications.
}
\description{
Train range for bivariate scale
}
