% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data_example.R
\docType{data}
\name{tree_data_example}
\alias{tree_data_example}
\title{tree_data_example}
\format{
A data frame with 14 rows and 4 columns representing the structure of trees:
\describe{
  \item{var}{Variable name used for splitting.}
  \item{value}{The value in a node (i.e., either the split value or leaf value).}
  \item{iteration}{Iteration Number.}
  \item{treeNum}{Tree Number in the iteration.}
  ...
}
}
\usage{
tree_data_example
}
\description{
Small example of tree data, like that obtained when using `extractTreeData()` function.
}
\keyword{datasets}
