% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBarPlot.R
\name{treeBarPlot}
\alias{treeBarPlot}
\title{treeBarPlot}
\usage{
treeBarPlot(
  treeData,
  iter = NULL,
  treeNo = NULL,
  topTrees = NULL,
  removeStump = FALSE,
  combineFact = FALSE
)
}
\arguments{
\item{treeData}{A list of tree attributes created using the extractTreeData function.}

\item{iter}{The selected iteration}

\item{treeNo}{The selected tree number.}

\item{topTrees}{integer value to show the top x variables.}

\item{removeStump}{LOGICAL. If TRUE, then stumps are removed from plot. If False, stumps
remain in plot and are coloured grey.}

\item{combineFact}{If a variable is a factor in a data frame, when building the BART model it is replaced with dummies.
Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor.
If combineFact = TRUE, then the dummy factors are combined into their original factor.}
}
\value{
A barplot plot.
}
\description{
Creates a barplot displaying the frequency of different tree structures.
}
