% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBarPlotOLD.R
\name{treeBarPlotOLD}
\alias{treeBarPlotOLD}
\title{treeBarPlotOLD}
\usage{
treeBarPlotOLD(treeList)
}
\arguments{
\item{treeList}{A list of trees created using the treeList function.}
}
\value{
A barplot plot.
}
\description{
Creates a barplot displaying the frequency of different tree structures.
}
