% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeData.R
\name{treeData}
\alias{treeData}
\title{treeData}
\usage{
treeData(model)
}
\arguments{
\item{model}{A model created from either the bart, dbarts, or bartMachine packages.}
}
\value{
A tibble of every tree and its attributes.
}
\description{
Creates a data frame of tree attributes for a model
created from either the bart, dbarts, or bartMachine packages.
}
