% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeList.R
\name{treeList}
\alias{treeList}
\title{treeList}
\usage{
treeList(trees)
}
\arguments{
\item{trees}{A data frame created by treeData function.}
}
\value{
A list of every tree and its attributes.
}
\description{
Creates a list of tree attributes for a model
created from either the bart or dbarts packages.
}
