% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap.R
\name{treeMap}
\alias{treeMap}
\title{treeMap}
\usage{
treeMap(treeList)
}
\arguments{
\item{treeList}{A list of trees created using the treeList function.}
}
\value{
A treemap plot.
}
\description{
Creates a treemap displaying the frequency of different tree structures.
}
