% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMapOLD.R
\name{treeMapOLD}
\alias{treeMapOLD}
\title{treeMapOLD}
\usage{
treeMapOLD(treeList)
}
\arguments{
\item{treeList}{A list of trees created using the treeList function.}
}
\value{
A treemap plot.
}
\description{
Creates a treemap displaying the frequency of different tree structures.
}
