% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpPlot.R
\name{vimpPlot}
\alias{vimpPlot}
\title{vimpPlot}
\usage{
vimpPlot(treeData, type = "prop", plotType = "barplot", metric = "median")
}
\arguments{
\item{treeData}{A data frame created by treeData function.}

\item{type}{What value to return. Either the raw count 'count'
or the proportions 'prop' averaged over iterations.}

\item{plotType}{Which type of plot to return. Either a barplot 'barplot' with the
quantiles shown as a line, a point plot with the quantiles shown as a gradient 'pointGrad', or a
letter-value plot 'lvp'.}

\item{metric}{Whether to show the 'mean' or 'median' importance values. Note, this has
no effect when using plotType = 'lvp'.}
}
\value{
A plot of variable importance.
}
\description{
Plot the variable importance for a BART model with the 25% and 75%
quantile.
}
