% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviData.R
\name{vintBart}
\alias{vintBart}
\title{vintBart}
\usage{
vintBart(treeData)
}
\arguments{
\item{treeData}{A data frame created by treeData function.}
}
\value{
A dataframe containing the count of each pair of variables and the mean of the proportions.
}
\description{
A data frame displaying the count of each pair of variables used
for splitting within a tree.
}
