% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviBartPlot.R
\name{viviBartPlot}
\alias{viviBartPlot}
\title{viviBartPlot}
\usage{
viviBartPlot(
  matrix,
  intPal = rev(colorspace::sequential_hcl(palette = "Purples 3", n = 2^4/2)),
  impPal = rev(colorspace::sequential_hcl(palette = "Greens 3", n = 2^4/2)),
  intLims = NULL,
  impLims = NULL,
  uncIntLims = NULL,
  uncImpLims = NULL,
  angle = 0,
  border = FALSE
)
}
\arguments{
\item{matrix}{Matrices, such as that returned by viviBartMatrix, of values to be plotted.}

\item{intPal}{A vector of colours to show interactions, for use with scale_fill_gradientn. Palette number has to be 2^x/2}

\item{impPal}{A vector of colours to show importance, for use with scale_fill_gradientn. Palette number has to be 2^x/2}

\item{intLims}{Specifies the fit range for the color map for interaction strength.}

\item{impLims}{Specifies the fit range for the color map for importance.}

\item{uncIntLims}{Specifies the fit range for the color map for interaction strength uncertainties.}

\item{uncImpLims}{Specifies the fit range for the color map for importance uncertainties.}

\item{angle}{The angle to rotate the x-axis labels. Defaults to zero.}

\item{border}{Logical. If TRUE then draw a black border around the diagonal elements.}
}
\value{
TBD
}
\description{
Plots a Heatmap showing variable importance on the diagonal
and variable interaction on the off-diagonal with uncertainty included.
}
