% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptenceRate.R
\name{acceptRate}
\alias{acceptRate}
\title{acceptRate}
\usage{
acceptRate(treeData, burnIn = 0)
}
\arguments{
\item{treeData}{A data frame created by treeData function.}

\item{burnIn}{Numerical value of the burn-in.
Displays a division on the plot to separate prior and post burn-in iterations.}
}
\value{
A plot of acceptence rate.
}
\description{
Plots the acceptence rate of trees from a BART model.
}
