% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptenceRate.R
\name{acceptRate}
\alias{acceptRate}
\title{acceptRate}
\usage{
acceptRate(treeData, burnIn = 0)
}
\arguments{
\item{treeData}{A data frame created by treeData function.
Displays a division on the plot to separate prior and post burn-in iterations.}
}
\value{
A plot of acceptance rate.
}
\description{
Plots the acceptance rate of trees from a BART model.
}
