% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptenceRate.R
\name{acceptRate}
\alias{acceptRate}
\title{acceptRate}
\usage{
acceptRate(trees)
}
\arguments{
\item{trees}{A data frame created by extractTreeData function.
Displays a division on the plot to separate prior and post burn-in iterations.}
}
\value{
A ggplot object plot of acceptance rate.
}
\description{
Plots the acceptance rate of trees from a BART model.
}
\examples{
\donttest{
df_trees <- extractTreeData(model = my_model, data = my_data)
acceptRate(trees = df_trees)
}


}
