% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartClassDiag.R
\name{bartClassDiag}
\alias{bartClassDiag}
\title{bartClassDiag}
\usage{
bartClassDiag(model, response)
}
\arguments{
\item{model}{a model created from either the BART, dbarts, or bartMachine package.}

\item{response}{The name of the response for the fit.}
}
\value{
A selection of diagnostic plots
}
\description{
Displays a selection of diagnostic plots for a BART model.
}
