% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartDiag.R
\name{bartDiag}
\alias{bartDiag}
\title{bartDiag}
\usage{
bartDiag(model, response, burnIn = 0, vImpReorder = FALSE)
}
\arguments{
\item{model}{a model created from the BART package}

\item{response}{The name of the response for the fit.}

\item{burnIn}{Trace plot will only show iterations above selected burn in value.}

\item{pal}{A vector of colours to show importance, for use with scale_fill_gradientn.}

\item{impLims}{Specifies the fit range for the color map for importance.}

\item{reorder}{If TRUE, then the variable importance plot is reordered in terms of importance.}
}
\value{
A selection of diagnostic plots
}
\description{
Displays a selection of diagnostic plots for a BART model.
}
