% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartDiag.R
\name{bartDiag}
\alias{bartDiag}
\title{bartDiag}
\usage{
bartDiag(
  model,
  data,
  response,
  burnIn = 0,
  threshold = "Youden",
  pNorm = FALSE,
  showInterval = TRUE,
  combineFactors = FALSE
)
}
\arguments{
\item{model}{a model created from either the BART, modelarts, or bartMachine package.}

\item{data}{A dataframe used to build the model.}

\item{response}{The name of the response for the fit.}

\item{burnIn}{Trace plot will only show iterations above selected burn in value.}

\item{threshold}{A dashed line on some plots to indicate a chosen threshold value (classification only).
by default the Youden index is shown.}

\item{pNorm}{apply pnorm to the y-hat data (classification only).}

\item{showInterval}{LOGICAL if TRUE then show 5\% and 95\% quantile intervals on ROC an PC curves (classification only).}

\item{combineFactors}{Whether or not to combine dummy variables (if present) in display.}
}
\value{
A selection of diagnostic plots
}
\description{
Displays a selection of diagnostic plots for a BART model.
}
