% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartDiag.R
\name{bartRegrDiag}
\alias{bartRegrDiag}
\title{bartRegrDiag}
\usage{
bartRegrDiag(model, response, burnIn = 0, data, combineFactors = FALSE)
}
\arguments{
\item{model}{a model created from either the BART, modelarts, or bartMachine package.}

\item{response}{The name of the response for the fit.}

\item{burnIn}{Trace plot will only show iterations above selected burn in value.}

\item{data}{A dataframe used to build the model.}

\item{combineFactors}{Whether or not to combine dummy variables (if present) in display.}
}
\value{
A selection of diagnostic plots
}
\description{
Displays a selection of diagnostic plots for a BART model.
}
