% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFactors.R
\name{combineFactorsDiag}
\alias{combineFactorsDiag}
\title{combineFactorsDiag}
\usage{
combineFactorsDiag(data, df2, model)
}
\arguments{
\item{data}{A data frame}

\item{df2}{A second data frame that has the factor split into individual levels.}

\item{model}{a BART model}
}
\description{
If a variable is a factor in a data frame, when building the BART model it is replaced with dummies.
Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor.
These functions combine the factor levels so that the inclusion proportions
are aggregated so the importance can be assessed for the the entire factor.
}
