% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTreeData.R
\name{extractTreeData}
\alias{extractTreeData}
\title{extractTreeData}
\usage{
extractTreeData(model)
}
\arguments{
\item{model}{Model created from either the BART, dbarts or bartMachine packages.}
}
\value{
A list of every tree and its attributes.
}
\description{
Creates a list of all tree attributes for a model
created by either the BART, dbarts or bartMachine packages.
}
