% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsBart.R
\name{mdsBart}
\alias{mdsBart}
\title{mdsBart}
\usage{
mdsBart(
  treeData,
  data,
  target,
  type = "mean",
  plotType = "rows",
  showGroup = TRUE
)
}
\arguments{
\item{treeData}{A data frame created by treeData function.}

\item{type}{What value to return. Either the raw count 'val', the proportion 'prop'
or the column means of the proportions 'propMean'}
}
\value{
For this function, the MDS coordinates are calculated for each iteration.
Procrustes method is then applied to align each of the coordinates to a target set
of coordinates. The returning result is then a clustered average of each point.
}
\description{
Multi-dimensional Scaling Plot of proximity matrix from a BART model.
}
