% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTrees.R
\name{plotAllTrees}
\alias{plotAllTrees}
\title{plotAllTrees}
\usage{
plotAllTrees(
  treeData,
  iter = NULL,
  treeNo = NULL,
  sampleSize = NULL,
  cluster = NULL,
  sizeNode = TRUE,
  pal = RColorBrewer::brewer.pal(9, "Purples"),
  fillBy = NULL,
  selectedVars = NULL,
  removeStump = FALSE,
  combineFact = FALSE
)
}
\arguments{
\item{treeData}{A list of tree attributes created by exctractTreeData function.}

\item{iter}{The selected iteration}

\item{treeNo}{The selected tree number.}

\item{sampleSize}{Sample the tree list.}

\item{cluster}{LOGICAL. If TRUE, then cluster by tree structures.}

\item{sizeNode}{Whether to size node width by the number of observations that fall into that node.}

\item{pal}{A palette to colour the terminal nodes.}

\item{fillBy}{Which parameter to colour the terminal nodes. Either 'response' or 'mu'.}

\item{removeStump}{LOGICAL. If TRUE, then stumps are removed from plot. If False, stumps
remain in plot and are coloured grey.}

\item{combineFact}{If a variable is a factor in a data frame, when building the BART model it is replaced with dummies.
Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor.
If combineFact = TRUE, then the dummy factors are combined into their original factor.}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
Plots all the trees. By default, it shows the last iteration. If number of
trees is greater than 200, a sample of trees from that iteration will be shown.
}
