% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTrees.R
\name{plotAllTrees}
\alias{plotAllTrees}
\title{plotAllTrees}
\usage{
plotAllTrees(
  treeData,
  iter = NULL,
  treeNo = NULL,
  sampleSize = NULL,
  cluster = NULL
)
}
\arguments{
\item{treeData}{A list of tree attributes created by treeDatafunction.}

\item{iter}{The selected iteration}

\item{treeNo}{The selected tree number.}

\item{sampleSize}{Sample the tree list.}

\item{cluster}{LOGICAL. If TRUE, then cluster by tree structures.}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
Plots all the trees. By default, it shows the last iteration. If number of
trees is greater than 200, a sample of trees from that iteration will be shown.
}
