% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTrees.R
\name{plotAllTreesPlotFn}
\alias{plotAllTreesPlotFn}
\title{plotAllTreesplotFn}
\usage{
plotAllTreesPlotFn(
  treeList,
  sizeNode = TRUE,
  pal = rev(colorRampPalette(c("steelblue", "#f7fcfd", "orange"))(5)),
  fillBy = NULL,
  selectedVars = NULL,
  removeStump = FALSE,
  combineFact = FALSE,
  treeData,
  name
)
}
\arguments{
\item{treeList}{A list of trees created by treeList function.}

\item{sizeNode}{Whether to size node width by the number of observations that fall into that node.}

\item{pal}{A palette to colour the terminal nodes.}

\item{fillBy}{Which parameter to colour the terminal nodes. Either 'response' or 'mu'.}

\item{selectedVars}{A vector of selected variables to display.}

\item{removeStump}{LOGICAL. If TRUE, then stumps are removed from plot. If False, stumps
remain in plot and are coloured grey.}

\item{combineFact}{If a variable is a factor in a data frame, when building the BART model it is replaced with dummies.
Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor.
If combineFact = TRUE, then the dummy factors are combined into their original factor.}

\item{treeData}{A list of tree attributes created by exctractTreeData function.}

\item{name}{Variable names}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
This function is used to creat the plots.
}
