% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTrees.R
\name{plotAllTreesPlotFn}
\alias{plotAllTreesPlotFn}
\title{plotAllTreesplotFn}
\usage{
plotAllTreesPlotFn(treeList, sampleSize = NULL)
}
\arguments{
\item{treeList}{A list of trees created by treeList function.}

\item{sampleSize}{Sample the tree list.}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
This function is used to creat the plots.
}
