% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTrees.R
\name{plotAllTreesPlotFn}
\alias{plotAllTreesPlotFn}
\title{plotAllTreesplotFn}
\usage{
plotAllTreesPlotFn(
  treeList,
  sampleSize = NULL,
  sizeNode = TRUE,
  pal = RColorBrewer::brewer.pal(9, "Purples")
)
}
\arguments{
\item{treeList}{A list of trees created by treeList function.}

\item{sampleSize}{Sample the tree list.}

\item{sizeNode}{Whether to size node width by the number of observations that fall into that node.}

\item{pal}{A palette to colour the terminal nodes.}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
This function is used to creat the plots.
}
