% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTrees.R
\name{plotAllTreesPlotFn}
\alias{plotAllTreesPlotFn}
\title{plotAllTreesplotFn}
\usage{
plotAllTreesPlotFn(
  treeList,
  sampleSize = NULL,
  sizeNode = TRUE,
  pal = rev(colorspace::sequential_hcl(palette = "Purples 3", n = 100)),
  fillBy = NULL,
  selectedVars = NULL,
  removeStump = FALSE,
  name
)
}
\arguments{
\item{treeList}{A list of trees created by treeList function.}

\item{sampleSize}{Sample the tree list.}

\item{sizeNode}{Whether to size node width by the number of observations that fall into that node.}

\item{pal}{A palette to colour the terminal nodes.}

\item{fillBy}{Which parameter to colour the terminal nodes. Either 'response' or 'mu'.}

\item{name}{Variable names}
}
\value{
A list containing vectors of the indices of observations from leaf nodes.
}
\description{
This function is used to creat the plots.
}
