% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleTree.R
\name{plotSingleTree}
\alias{plotSingleTree}
\title{plotSingleTree}
\usage{
plotSingleTree(trees, iter = 1, treeNo = 1, plotType = "icicle")
}
\arguments{
\item{trees}{A data frame created by \code{extractTreeData} function}

\item{iter}{The MCMC iteration or chain to plot.}

\item{treeNo}{The tree number to plot.}

\item{plotType}{What type of plot to display. either dendrogram or icicle.}
}
\value{
A plot of an individual tree
}
\description{
Plots individual trees.
}
\examples{
\dontrun{
df_trees <- extractTreeData(model = my_model, data = my_data)
plotSingleTree(trees = df_trees, treeNo = 1, iter = 1, plotType = "icicle")
}
}
