% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTrees}
\alias{plotTrees}
\title{plotTrees}
\usage{
plotTrees(
  model,
  data,
  response,
  treeNum = 1,
  iteration = 1,
  plotType = c("dendrogram", "icicle")
)
}
\arguments{
\item{model}{A bart model created from he R-packages BART, dbarts, or
bartMachine.}

\item{data}{Data frame used for fit.}

\item{response}{The name of the response for the fit.}

\item{treeNum}{The tree number to plot.}

\item{plotType}{What type of plot to display. either dendrogram or icicle.}

\item{iter}{The MCMC iteration or chain to plot.}
}
\value{
A plot of an individual tree
}
\description{
plots individual trees from the R-packages BART, dbarts, and
bartMachine
}
