% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTrees}
\alias{plotTrees}
\title{plotTrees}
\usage{
plotTrees(
  treeData,
  treeNum = 1,
  iteration = 1,
  plotType = c("dendrogram", "icicle")
)
}
\arguments{
\item{treeData}{A data frame created by treeList function}

\item{treeNum}{The tree number to plot.}

\item{iteration}{The MCMC iteration or chain to plot.}

\item{plotType}{What type of plot to display. either dendrogram or icicle.}
}
\value{
A plot of an individual tree
}
\description{
plots individual trees from the R-packages BART, dbarts, and
bartMachine
}
