% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTreeData.R
\name{print.hideHelper1}
\alias{print.hideHelper1}
\title{print.hideHelper}
\usage{
\method{print}{hideHelper1}(x, ...)
}
\arguments{
\item{x}{A data frame of trees}

\item{...}{Extra parameters}
}
\value{
No return value; this function is called for its side effect of printing a formatted summary
of the tree data frame. It displays parts of the data frame, such as the tree structure and various
counts (like number of MCMC iterations, number of trees, and number of variables), while keeping
the complete data accessible via indexing.
}
\description{
This function hides parts from the print out
but are still accessible via indexing.
}
