% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitymatrix.R
\name{proximityMatrix}
\alias{proximityMatrix}
\title{proximityMatrix}
\usage{
proximityMatrix(treeData, nRows, normalize = TRUE, reorder = TRUE, iter = NULL)
}
\arguments{
\item{treeData}{A list of tree attributes created by treeData function.}

\item{nRows}{Number of rows to consider.}

\item{normalize}{Default is TRUE. Divide the total number of pairs of observations by
the number of trees.}

\item{reorder}{Default is TRUE. Whether to sort the matrix so high values are pushed to top left.}

\item{iter}{Which iteration to use, if NULL the proximity matrix is calculated over all
iterations.}
}
\value{
A matrix containing proximity values.
}
\description{
Creates a matrix of proximity values.
}
