% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDensity.R
\name{splitDensity}
\alias{splitDensity}
\title{splitDensity}
\usage{
splitDensity(treeData, data, display = "histogram")
}
\arguments{
\item{treeData}{A list of trees created using the treeData function.}

\item{data}{Data frame containing variables from the model.}

\item{display}{Choose how to display the plot. Either histogram, facet wrap, ridges
or display both the split value and density of the predictor by using dataSplit.}
}
\value{
A faceted group of density plots
}
\description{
Density plots of the split value for each variable.
}
