% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBarPlot.R
\name{treeBarPlot}
\alias{treeBarPlot}
\title{treeBarPlot}
\usage{
treeBarPlot(treeList)
}
\arguments{
\item{treeList}{A list of trees created using the treeList function.}
}
\value{
A barplot plot.
}
\description{
Creates a barplot displaying the frequency of different tree structures.
}
