% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBarPlot.R
\name{treeBarPlot}
\alias{treeBarPlot}
\title{treeBarPlot}
\usage{
treeBarPlot(
  treeData,
  iter = NULL,
  treeNo = NULL,
  topTrees = NULL,
  removeStump = FALSE
)
}
\arguments{
\item{treeData}{A list of tree attributes created using the extractTreeData function.}

\item{iter}{The selected iteration}

\item{treeNo}{The selected tree number.}

\item{topTrees}{integer value to show the top x variables.}

\item{removeStump}{LOGICAL. If TRUE, then stumps are removed from plot. If False, stumps
remain in plot and are coloured grey.}
}
\value{
A barplot plot.
}
\description{
Creates a barplot displaying the frequency of different tree structures.
}
