% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDepth.R
\name{treeDepth}
\alias{treeDepth}
\title{treeDepth}
\usage{
treeDepth(trees)
}
\arguments{
\item{trees}{A list of tree attributes created using the extractTreeData function.}
}
\value{
A plot of average tree depths over iteration
}
\description{
A plot of tree depth over iterations.
}
\examples{
\donttest{
df_trees <- extractTreeData(model = my_model, data = my_data)
treeDepth(trees = df_trees)
}

}
