% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDepth.R
\name{treeDepth}
\alias{treeDepth}
\title{treeDepth}
\usage{
treeDepth(treeData, burnIn)
}
\arguments{
\item{treeData}{A list of tree attributes created using the extractTreeData function.}

\item{burnIn}{Selected burn in value.}
}
\value{
A plot of average tree depths over iteration
}
\description{
A plot of tree depth over iterations.
}
