% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeNodes.R
\name{treeNodes}
\alias{treeNodes}
\title{treeNodes}
\usage{
treeNodes(trees)
}
\arguments{
\item{trees}{A list of tree attributes created using the extractTreeData function.}
}
\value{
A plot of tree number of nodes over iterations.
}
\description{
A plot of number of nodes over iterations.
}
\examples{
\donttest{
df_trees <- extractTreeData(model = my_model, data = my_data)
treeNodes(trees = df_trees)
}

}
