% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeNodes.R
\name{treeNodes}
\alias{treeNodes}
\title{treeNodes}
\usage{
treeNodes(treeData)
}
\arguments{
\item{treeData}{A list of tree attributes created using the extractTreeData function.}
}
\value{
A plot of tree number of nodes over iterations.
}
\description{
A plot of number of nodes over iterations.
}
