% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpBart.R
\name{vimpBart}
\alias{vimpBart}
\title{vimpBart}
\usage{
vimpBart(treeData, type = "prop")
}
\arguments{
\item{treeData}{A data frame created by treeData function.}

\item{type}{What value to return. Either the raw count 'val', the proportion 'prop',
the column means of the proportions 'propMean', or the median of the proportions 'propMedian'.}
}
\value{
A matrix
}
\description{
A matrix with nMCMC rows with each variable as a column.
Each row represents an MCMC iteration. For each variable, the total count
of the number of times that variable is used in a tree is given.
}
