% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintPlot.R
\name{vintPlot}
\alias{vintPlot}
\title{vintPlot}
\usage{
vintPlot(treeData, plotType = "barplot", top = NULL)
}
\arguments{
\item{treeData}{A data frame created by treeData function.}

\item{plotType}{Which type of plot to return. Either a barplot 'barplot' with the
quantiles shown as a line, a point plot with the quantiles shown as a gradient 'point', or a
letter-value plot 'lvp'.}

\item{top}{Display only the top X metrics (does not apply to the letter-value plot).}
}
\value{
A plot of variable importance.
}
\description{
Plot the pair-wise variable interactions inclusion porportions
for a BART model with the 25% and 75% quantile.
}
