% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviBART.R
\name{viviBART}
\alias{viviBART}
\title{viviBART}
\usage{
viviBART(
  model,
  response,
  data,
  noReplications = 2,
  gridSize = 50,
  nmax = 500,
  reorder = TRUE,
  class = 1,
  normalized = FALSE
)
}
\arguments{
\item{response}{The name of the response for the fit.}

\item{data}{Data frame used for fit.}

\item{noReplications}{How many time to repeat the calculation.}

\item{gridSize}{The size of the grid for evaluating the predictions.}

\item{nmax}{Maximum number of data rows to consider. Default is 500. Use all rows if NULL.}

\item{reorder}{If TRUE (default) uses DendSer to reorder the matrix of interactions and variable importances.}

\item{class}{Category for classification, a factor level, or a number indicating which factor level.}

\item{normalized}{Should Friedman's H-statistic be normalized or not. Default is FALSE.}

\item{fit}{A supervised machine learning model, which understands condvis2::CVpredict}
}
\value{
A list of matrices. One matrix is of interaction values with importance on the diagonal. The second
is a matrix of their uncertainties.
}
\description{
Creates a matrix displaying variable importance on the diagonal
 and variable interaction on the off-diagonal with the uncertainty included.
}
\examples{

aq <- na.omit(airquality)
f <- lm(Ozone ~ ., data = aq)
m <- vivi(fit = f, data = aq, response = "Ozone") # as expected all interactions are zero
viviHeatmap(m)

\donttest{
library(ranger)
rf <- ranger(Species ~ ., data = iris, importance = "impurity", probability = TRUE)
vivi(fit = rf, data = iris, response = "Species")
}
}
