% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviBart.R
\name{viviBart}
\alias{viviBart}
\title{viviBart}
\usage{
viviBart(treeData)
}
\arguments{
\item{treeData}{A data frame created by extractTreeData function.}
}
\value{
A list of dataframes of VIVI summaries.
}
\description{
Returns a list containing a dataframe of variable importance summaries
and a dataframe of variable interaction summaries.
}
