% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviBartMatrix.R
\name{viviBartMatrix}
\alias{viviBartMatrix}
\title{viviBartMatrix}
\usage{
viviBartMatrix(
  treeData,
  type = "standard",
  metric = "propMean",
  metricError = "SD",
  reorder = FALSE
)
}
\value{
A heatmap plot showing variable importance on the diagonal
and variable interaction on the off-diagonal.
}
\description{
Returns a matrix or list of matrices. If type = 'standard' a
matrix filled with vivi values is returned. If type = 'vsup' two matrices are returned.
One with the actual values and another matrix of uncertainty values.
If type = 'quantiles, three matrices are returned. One for the 25%, 50%, and 75% quantiles.

 @param treeData A data frame created by extractTreeData function.
 @param type Which type of matrix to return. Either 'standard', 'vsup', 'quantiles'
 @param metric Which metric to use to fill the actual values matrix. Either 'propMean' or 'count'.
 @param metricError Which metric to use to fill the uncertainty matrix. Either 'SD' or 'SEofCI'.
 @param reorder LOGICAL. If TRUE then the matrix is reordered so high values are pushed to the top left.
}
