% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviBART.R
\name{viviBartPlot}
\alias{viviBartPlot}
\title{viviBartPlot}
\usage{
viviBartPlot(
  matrix,
  intPal = rev(colorspace::sequential_hcl(palette = "Purples 3", n = 2^4/2)),
  impPal = rev(colorspace::sequential_hcl(palette = "Greens 3", n = 2^4/2)),
  intLims = NULL,
  impLims = NULL,
  uncIntLims = NULL,
  uncImpLims = NULL,
  border = FALSE
)
}
\arguments{
\item{intPal}{A vector of colours to show interactions, for use with scale_fill_gradientn.}

\item{impPal}{A vector of colours to show importance, for use with scale_fill_gradientn.}

\item{intLims}{Specifies the fit range for the color map for interaction strength.}

\item{impLims}{Specifies the fit range for the color map for importance.}

\item{border}{Logical. If TRUE then draw a black border around the diagonal elements.}

\item{mat}{Matrices, such as that returned by viviBART, of values to be plotted.}

\item{angle}{The angle to rotate the x-axis labels. Defaults to zero.}
}
\value{
A heatmap plot showing variable importance on the diagonal
and variable interaction on the off-diagonal.
}
\description{
Plots a Heatmap showing variable importance on the diagonal
and variable interaction on the off-diagonal with uncertainty included.
}
